backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
aisqlite==1.1.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bitarray==3.4.3
blinker==1.6
Brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
cryptography==45.0.5
cytoolz==1.0.1
distro==1.9.0
docutils==0.21.2
easyocr==1.7.2
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
fastapi==0.115.14
filelock==3.13.1
frozenlist==1.7.0
fsspec==2024.6.1
greenlet==3.2.3
h11==0.16.0
h2==4.2.0
hexbytes==1.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
id==1.5.0
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
Jinja2==3.1.4
jiter==0.10.0
kaitaistruct==0.10
keyring==25.6.0
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
more-itertools==10.7.0
mpmath==1.3.0
multidict==6.6.3
networkx==3.3
nh3==0.2.21
ninja==1.11.1.4
numpy<2
openai==1.93.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==25.0
parsimonious==0.10.0
pdf2image==1.17.0
pillow==11.3.0
pluggy==1.6.0
propcache==0.3.2
pyasn1==0.6.1
pyclipper==1.3.0.post6
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyOpenSSL==25.1.0
pyparsing==3.2.3
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.4.1
pytest-asyncio==1.0.0
python-bidi==0.6.6
python-dotenv==1.1.1
python-multipart==0.0.20
pyunormalize==16.0.0
PyYAML==6.0.2
readme_renderer==44.0
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
rlp==4.1.0
scikit-image==0.25.2
scipy==1.16.0
selenium==4.34.0
selenium-wire==5.1.0
setuptools==80.9.0
shapely==2.1.1
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.13.3
tifffile==2025.6.11
toolz==1.0.0
torch==2.2.2
torchaudio==2.2.2
torchvision==0.17.2
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
twine==6.1.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.35.0
web3==7.12.0
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
wsproto==1.2.0
yarl==1.20.1
zstandard==0.23.0